Calculate quality of sum score of selected variables

me_sscore takes a data frame with quality estimates of class me and estimates the quality of a sum score for selected variables in .... Use me_sscore_ if you're interested in programming with me_sscore

me_sscore(me_data, .data, new_name, ..., wt = NULL, drop = TRUE)

me_sscore_(me_data, .data, new_name, vars_names, wt = NULL,
  drop = TRUE)

Arguments

me_data
a data frame of class me containing quality estimates from the variables specified in ....

.data
a data frame which contains data for the variables specified in ....

new_name
a bare unquoted name or a string specifying the name of the new sum score.

...
bare unquoted names or separate strings specifying the variable names from which to estimate quality of their sum score. They all must be present in me_data and df. At minimum, it must be two or more variable names.

wt
a non-NA numeric vector of the same length as the variables specified in .... This will be used as weights in calculating the sum scores of all variable. Be default, all variables are given the same weight.

drop
a logical stating whether to drop the questions that compose the sum score (specified in ...) If FALSE it retains the original questions and the composite score.

vars_names
character vector specifying the variable names from which to estimate quality of their sum score. They all must be present in me_data and df. At minimum, it must be two or more variable names.

Value

a tibble similar to me_data but with a new row containing the quality of a sum score with the name specified in new_name.

Examples

# Toy data

library(tibble)
me_df <-
tibble(question = paste0("V", 1:5),
       quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
       reliability = c(0.2, 0.4, 0.5, 0.5, 0.7),
       validity = c(0.8, 0.1, 0.6, 0.7, 0.8))

sample_data <-
 as_tibble(
 stats::setNames(
  replicate(5, rbinom(1000, 5, 0.6), simplify = FALSE),
 paste0("V", 1:5))
 )

me_sscore(
me_data = me_df,
.data = sample_data,
new_name = new_sumscore,
V3, V4
)

#> # A tibble: 4 x 4
#>   question     reliability validity quality
#>   <chr>              <dbl>    <dbl>   <dbl>
#> 1 V1                   0.2      0.8   0.2  
#> 2 V2                   0.4      0.1   0.3  
#> 3 V5                   0.7      0.8   0.9  
#> 4 new_sumscore        NA       NA     0.410

me_sscore(
me_data = me_df,
.data = sample_data,
new_name = new_sumscore,
"V1", "V2"
)

#> # A tibble: 4 x 4
#>   question     reliability validity quality
#>   <chr>              <dbl>    <dbl>   <dbl>
#> 1 V3                   0.5      0.6   0.5  
#> 2 V4                   0.5      0.7   0.6  
#> 3 V5                   0.7      0.8   0.9  
#> 4 new_sumscore        NA       NA     0.149