Define your measurement error design

The model syntax allows to define the relationship between variables in terms of whether they share a common method in the question that was asked. For example, if the questions called stflife and stfwork were both measured with a 1-10 likert scale related to satisfaction, these two questions were measured with the same method. The model syntax would specify them like '~ stflife + stfwork'.

medesign(model_syntax, .data, me_data, ...)

Arguments

model_syntax
A description of the user-specified model as a character string. The model is described using the me model syntax. See the details and example section for on how to write this syntax.

.data
a data frame containing all variables defined in model_syntax

me_data
A dataframe of class me. Can be created manually with me_construct.

...
arguments passed to me_sscore.

Value

an me design object which can be passed to other functions of measurementfree.

Details

These two questions should also be in .data and in me_data, otherwise no calculation can be performed. .data should be a data frame with all variables specified in the model_syntax while me_data should be a dataframe with 4 columns: question, reliability, validity an quality. The first column should contain the names of the variables as character vector, while the other three should be numeric vectors with their corresponding reliability, validity and quality values.

Currently, medesign replaces the non-NA quality of all variables found in me_data and .data in the correlation/covariance diagonal by default. This is then used by me_cmv_cor and me_cmv_cov to correct for the quality in the diagonal.

Examples

# These three variables share a common method
me_syntax <- "~ mpg + cyl + drat"

# Fake data for the example
me_data <- data.frame(stringsAsFactors = FALSE,
                      question = c("mpg", "cyl", "drat"),
                      reliability = c(0.729, 0.815, 0.68),
                      validity = c(0.951, 0.944, 0.79),
                      quality = c(0.693, 0.77, 0.89)
                      )

medesign(me_syntax, mtcars, me_data)

#> <Measurement error design>
#> Parsed model:
#>    ~ mpg + cyl + drat

# Each pair of variables share a common method.
me_syntax <-
"# Let's assume these two variables share a common method
 ~ mpg + cyl
 # Let's assume these two variables as well
 ~ drat + disp"

# Fake data for the example
me_data <- data.frame(stringsAsFactors = FALSE,
                      question = c("mpg", "cyl", "drat", "disp"),
                      reliability = c(0.729, 0.815, 0.68, 0.78),
                      validity = c(0.951, 0.944, 0.79, 0.97),
                      quality = c(0.693, 0.77, 0.89, 0.68)
                      )

medesign(me_syntax, mtcars, me_data)

#> <Measurement error design>
#> Parsed model:
#>    ~ mpg + cyl
#>    ~ drat + disp